restructured APIs
