+++ routes/addInterest.js
@@ -1,0 +1,37 @@
+let mysql = require('mysql');
+let config = require('../config.js');
+const fetch = require('node-fetch');
+const express = require("express");
+const path = require("path");
+const bodyParser = require("body-parser");
+
+const { response } = require('express');
+const router = express.Router();
+const port = process.env.PORT || 5000;
+router.use(bodyParser.json({ limit: '50mb' }));
+router.use(bodyParser.urlencoded({ limit: '50mb', extended: true }));
+
+router.use(express.static(path.join(__dirname, "client/build")));
+
+// // ---------------------------------------------------------------------
+
+app.post('/api/addInterest', (req, res) => {
+
+	let connection = mysql.createConnection(config);
+
+	let sql = `INSERT INTO osellner.Interested (renter_id, posting_id) VALUES (?, ?)`;
+	let data = [req.body.renter_id, req.body.posting_id];
+	
+	connection.query(sql, data, (error, results, fields) => {
+		if (error) {
+			return console.error(error.message);
+		}
+
+		let string = JSON.stringify(results);
+		// let obj = JSON.parse(string);
+		res.send({ express: string });
+	});
+	connection.end();
+});
+
+module.exports = router;
