+++ routes/getRenterUserID.js
@@ -1,0 +1,39 @@
+let mysql = require('mysql');
+let config = require('../config.js');
+const fetch = require('node-fetch');
+const express = require("express");
+const path = require("path");
+const bodyParser = require("body-parser");
+
+const { response } = require('express');
+const router = express.Router();
+const port = process.env.PORT || 5000;
+router.use(bodyParser.json({ limit: '50mb' }));
+router.use(bodyParser.urlencoded({ limit: '50mb', extended: true }));
+
+router.use(express.static(path.join(__dirname, "client/build")));
+
+// // ---------------------------------------------------------------------
+
+app.post('/api/getRenterUserID', (req, res) => {
+
+	let connection = mysql.createConnection(config);
+
+	let sql = `SELECT renter_id FROM osellner.Renters
+	WHERE osellner.Renters.email LIKE ?
+	`;
+	console.log(sql);
+	let data = [req.body.email];
+
+	connection.query(sql, data, (error, results, fields) => {
+		if (error) {
+			return console.error(error.message);
+		}
+
+		let obj = JSON.stringify(results);
+		res.send({ express: obj });
+	});
+	connection.end();
+});
+
+module.exports = router;
